+++ lib/src/visitor/async_evaluate.dart
@@ -1645,6 +1645,8 @@ class _EvaluateVisitor
       }
     } on SassException catch (error, stackTrace) {
       throwWithTrace(_exception(error.message, error.span), stackTrace);
+    } on ArgumentError catch (error, stackTrace) {
+      throwWithTrace(_exception(error.toString()), stackTrace);
     } catch (error, stackTrace) {
       String? message;
       try {